backend/requirements.txt
fastapi
uvicorn[standard]
aiogram
python-dotenv
python-jose
passlib[bcrypt]
requests
pydantic
python-multipart
